Move functions to separate module
